Added a few more small leaftasks. Got the craftItem prebuilt task working. Now just need to worry about the stalled crafting jobs. And also displaying all this on the UI....

+++ core/src/com/mygdx/game/util/DataBuilder.java
@@ -733,7 +733,7 @@ public class DataBuilder implements IDestroyable{
         public String name, displayName;
         public String[] items;
         public int[] itemAmounts;
-        public int time=0;
+        public float time=0;
     }
 
     /**
